BREAKING
Now LDC on Windows must be properly configured and will otherwise refuse to link a plugin.

Stop #328 workaround, in the name of dub.json simplifying.
The problem is that it only works for dplug-build, and is otherwise disrespectful to software configuration of LDC (which supports many methods).
We recommend that instead, you configure your etc/ldc.conf on Windows so that you get fast builds. See the Wiki "Getting Started".

If you really want LDC auto-detection of VS, create an envvar LDC_VSDIR with a value of "a-non-existing-path"